refactor: extract routes (#57)
* fix: extract routes

* fix: mainActivity develop

* fix: moving routes and objects

* chore: optimize imports

* chore: using string resource and optimize routes
