+++ app/src/main/java/io/github/sustainow/presentation/ui/ExpectedEnergyScreen.kt
@@ -1,5 +1,6 @@
 package io.github.sustainow.presentation.ui
 
+import ConsumptionMainPage
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -22,7 +23,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import io.github.sustainow.ConsumptionMainPage
 import io.github.sustainow.R
 import io.github.sustainow.domain.model.Question
 import io.github.sustainow.domain.model.UserState
