+++ app/src/main/java/io/github/sustainow/routes/CollectiveActionsRoutes.kt
@@ -1,0 +1,20 @@
+import kotlinx.serialization.Serializable
+
+@Serializable
+object CreateCollectiveAction
+
+@Serializable
+data class UpdateCollectiveAction(
+    val id: Int?,
+)
+
+@Serializable
+object CollectiveActions
+
+@Serializable
+object SearchCollectiveActions
+
+@Serializable
+data class ViewCollectiveAction(
+    val id: Int?,
+)
