+++ app/src/main/java/io/github/sustainow/routes/ConsumptionRoutes.kt
@@ -1,0 +1,22 @@
+import kotlinx.serialization.Serializable
+
+@Serializable
+object Consume
+
+@Serializable
+object ConsumptionMainPage
+
+@Serializable
+object ExpectedEnergyConsumption
+
+@Serializable
+object ExpectedWaterConsumption
+
+@Serializable
+object ExpectedCarbonFootprint
+
+@Serializable
+object RealEnergyConsumption
+
+@Serializable
+object RealWaterConsumption
