+++ app/src/main/java/io/github/sustainow/routes/GeneralRoutes.kt
@@ -1,0 +1,22 @@
+import kotlinx.serialization.Serializable
+
+@Serializable
+object Home
+
+@Serializable
+object Authentication
+
+@Serializable
+object Login
+
+@Serializable
+object SignUp
+
+@Serializable
+object Routines
+
+@Serializable
+object ViewRoutine
+
+@Serializable
+object Configuration
