+++ app/src/main/java/io/github/sustainow/routes/HistoricRoutes.kt
@@ -1,0 +1,18 @@
+package io.github.sustainow.routes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+object Historic
+
+@Serializable
+object HistoricMainPage
+
+@Serializable
+object HistoricConsumeWater
+
+@Serializable
+object HistoricConsumeEnergy
+
+@Serializable
+object HistoricCarbonFootprint
